добавил шаг с открытием браузера и убрал промежуточный класс implementation
