+++ src/test/resources/org/example/mainPageTests.feature
@@ -3,7 +3,8 @@
 Функционал: проверка
 
   Предыстория:
-    Когда Открываю страницу 'MAIN_PAGE'
+    Когда Открываю браузер 'chrome'
+    И Открываю страницу 'MAIN_PAGE'
     И Нажимаю на кнопку 'ОК'
 
   Сценарий: Проверка выбора курса по названию
